requirements.txt
numpy
ursina
numpy-quaternion